bookstore_app/pyfiles/requirements.txt
flask
mysql-connector-python
k 